[cleanup] Move some misc methods to empty utils module
